refactor(model): app_config -> config にディレクトリ名を変更
