+++ app/mobile/lib/flavor/flavor.dart
@@ -1,4 +1,4 @@
-import 'package:core_model/app_config.dart';
+import 'package:core_model/config.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:mobile_app/flavor/dev/firebase_options.dart' as dev;
 import 'package:mobile_app/flavor/prod//firebase_options.dart' as prod;
