+++ app/mobile/lib/initializer/firebase_initializer.dart
@@ -1,4 +1,4 @@
-import 'package:core_model/app_config.dart';
+import 'package:core_model/config.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
