--- core/model/lib/app_config.dart
@@ -1,1 +1,0 @@
-export 'src/app_config/index.dart';
