+++ core/model/lib/config.dart
@@ -1,0 +1,1 @@
+export 'src/config/index.dart';
