+++ core/model/lib/src/config/app_config.freezed.dart
