+++ core/model/lib/src/config/flavor.dart
